+++ rotkehlchen/api/v1/schemas.py
@@ -900,12 +900,19 @@ class CreateHistoryEventSchema(Schema):
             ),
         )
 
+        def __init__(self, dbhandler: 'DBHandler') -> None:
+            super().__init__()
+            self.database = dbhandler
+
         @post_load
         def make_history_base_entry(
                 self,
                 data: dict[str, Any],
                 **_kwargs: Any,
         ) -> dict[str, Any]:
+            with self.database.conn.read_ctx() as cursor:
+                tracked_accounts = self.database.get_blockchain_accounts(cursor)
+            data['fee_recipient_tracked'] = data['fee_recipient'] in tracked_accounts.get(SupportedBlockchain.ETHEREUM)  # noqa: E501
             return {'events': [EthBlockEvent(**data)]}
 
     class CreateEthDepositEventEventSchema(BaseSchema):